Refactoring file structure, adding better comments
