+++ src/api/index.ts
@@ -1,9 +1,11 @@
 import Jiter from '@jiter/node';
 import { Router } from 'express';
 import { events } from './events';
-import { jiterWebhookEvent } from './webhooks';
+import { webhooks } from './webhooks';
 
 export const api = Router();
 
+// Mount two additional routers
+// Each router handles different methods and may incorporate more nested path routers
 api.use('/events', events);
-api.use('/webhooks/jiter', jiterWebhookEvent);
+api.use('/webhooks', webhooks);
