+++ src/api/webhooks/index.ts
@@ -1,0 +1,6 @@
+import { Router } from 'express';
+import { groceriesHandler } from './groceriesHandler';
+
+export const webhooks = Router();
+
+webhooks.use('/groceries', groceriesHandler);
